Field longitud is now integer

+++ db/funcions/create-table-inventario-carretera-municipio.sql
@@ -6,7 +6,7 @@ CREATE TABLE inventario.carretera_municipio(
        orden_tramo varchar(1),
        pk_inicial_tramo float,
        pk_final_tramo float,
-       longitud_tramo float,
+       longitud_tramo integer,
        observaciones_tramo text,
        CONSTRAINT pk_carretera_municipio PRIMARY KEY (codigo_municipio, codigo_carretera, orden_tramo),
        FOREIGN KEY (codigo_carretera) REFERENCES inventario.carreteras (numero) ON DELETE CASCADE
